Refactor import formatting and update interface paths
Standardize spacing in import statements across files for consistency. Replace paths for interface imports to align with updated file structure. Minor tweaks in template code to enhance readability and maintain uniformity.
