+++ src/app/auth/auth.component.spec.ts
@@ -1,6 +1,6 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import {ComponentFixture, TestBed} from '@angular/core/testing';
 
-import { AuthComponent } from './auth.component';
+import {AuthComponent} from './auth.component';
 
 describe('AuthComponent', () => {
   let component: AuthComponent;
@@ -10,7 +10,7 @@ describe('AuthComponent', () => {
     await TestBed.configureTestingModule({
       imports: [AuthComponent]
     })
-    .compileComponents();
+      .compileComponents();
 
     fixture = TestBed.createComponent(AuthComponent);
     component = fixture.componentInstance;
