+++ src/app/auth/auth.component.ts
@@ -1,29 +1,28 @@
-import { NgClass, NgOptimizedImage } from "@angular/common";
-import { HttpErrorResponse } from "@angular/common/http";
-import { Component, inject } from '@angular/core';
-import { FormControl, FormGroup, FormsModule, ReactiveFormsModule, Validators } from "@angular/forms";
-import { Router } from "@angular/router";
-import { MessageService } from "primeng/api";
-import { Button } from "primeng/button";
-import { CardModule } from "primeng/card";
-import { DividerModule } from "primeng/divider";
-import { FloatLabelModule } from "primeng/floatlabel";
-import { IconFieldModule } from "primeng/iconfield";
-import { InputIconModule } from "primeng/inputicon";
-import { InputTextModule } from "primeng/inputtext";
-import { SplitterModule } from "primeng/splitter";
-import { StyleClassModule } from "primeng/styleclass";
-import { ToastModule } from "primeng/toast";
-import { AuthenticationService } from "./authentication.service";
-import { JWTInformation } from "../core/interfaces/JWTInformation";
-import { MessageDTO } from "../core/interfaces/MessageDTO";
+import {NgOptimizedImage} from "@angular/common";
+import {HttpErrorResponse} from "@angular/common/http";
+import {Component, inject} from '@angular/core';
+import {FormControl, FormGroup, FormsModule, ReactiveFormsModule, Validators} from "@angular/forms";
+import {Router} from "@angular/router";
+import {MessageService} from "primeng/api";
+import {Button} from "primeng/button";
+import {CardModule} from "primeng/card";
+import {DividerModule} from "primeng/divider";
+import {FloatLabelModule} from "primeng/floatlabel";
+import {IconFieldModule} from "primeng/iconfield";
+import {InputIconModule} from "primeng/inputicon";
+import {InputTextModule} from "primeng/inputtext";
+import {SplitterModule} from "primeng/splitter";
+import {StyleClassModule} from "primeng/styleclass";
+import {ToastModule} from "primeng/toast";
+import {AuthenticationService} from "./authentication.service";
+import {JWTInformation} from "../core/types/other/JWTInformation";
+import {MessageDTO} from "../core/types/other/MessageDTO";
 
 @Component({
   selector: 'app-auth',
   standalone: true,
   imports: [
     CardModule,
-    NgClass,
     StyleClassModule,
     NgOptimizedImage,
     IconFieldModule,
@@ -45,15 +44,14 @@ import { MessageDTO } from "../core/interfaces/MessageDTO";
 })
 export class AuthComponent {
 
-  // Injections
-  private readonly router = inject(Router);
-  private readonly messageService = inject(MessageService);
-  private readonly authenticationService = inject(AuthenticationService);
-
   loginForm: FormGroup = new FormGroup({
     username: new FormControl("", Validators.required),
     password: new FormControl("", Validators.required)
   });
+  // Injections
+  private readonly router = inject(Router);
+  private readonly messageService = inject(MessageService);
+  private readonly authenticationService = inject(AuthenticationService);
 
   public login() {
     if (this.loginForm.valid) {
