+++ src/app/auth/authentication.service.spec.ts
@@ -1,6 +1,6 @@
-import { TestBed } from '@angular/core/testing';
+import {TestBed} from '@angular/core/testing';
 
-import { AuthenticationService } from './authentication.service';
+import {AuthenticationService} from './authentication.service';
 
 describe('AuthenticationService', () => {
   let service: AuthenticationService;
